+++ app/src/main/java/com/philipgreen/dmwizard/player/PlayerCharacter.java
@@ -4,7 +4,7 @@ import android.util.Log;
 
 import com.philipgreen.dmwizard.data.Alignment;
 import com.philipgreen.dmwizard.data.BaseStats;
-import com.philipgreen.dmwizard.data.Languages;
+import com.philipgreen.dmwizard.player.utils.Languages;
 import com.philipgreen.dmwizard.player.utils.Skills;
 import com.philipgreen.dmwizard.data.Weapons;
 import com.philipgreen.dmwizard.player.utils.AbilityModifierManager;
